+++ test/browser/index.jsx
@@ -39,6 +39,9 @@ describe('Typeahead', () => {
         render(<Typeahead />, scratch)
 
         expect(scratch.innerHTML).to.contain('input')
+        expect(scratch.innerHTML).to.contain('class="typeahead__input"')
+        expect(scratch.innerHTML).to.contain('class="typeahead__menu"')
+        expect(scratch.innerHTML).to.contain('name="input-typeahead"')
       })
 
       it('renders an input with a name attribute', () => {
@@ -46,16 +49,30 @@ describe('Typeahead', () => {
 
         expect(scratch.innerHTML).to.contain('name="bob"')
       })
+
+      it('renders an input with a custom CSS namespace', () => {
+        render(<Typeahead cssNamespace='bob' />, scratch)
+
+        expect(scratch.innerHTML).to.contain('class="bob__input"')
+        expect(scratch.innerHTML).to.contain('class="bob__menu"')
+      })
     })
   })
 
   describe('behaviour', () => {
-    let typeahead
+    let typeahead, autoselectTypeahead
 
     beforeEach(() => {
       typeahead = new Typeahead({
         ...Typeahead.defaultProps,
         id: 'test',
+        source: suggest
+      })
+
+      autoselectTypeahead = new Typeahead({
+        ...Typeahead.defaultProps,
+        autoselect: true,
+        id: 'test2',
         source: suggest
       })
     })
@@ -108,23 +125,23 @@ describe('Typeahead', () => {
         typeahead.setState({ query: 'f' })
         typeahead.handleInputFocus()
         expect(typeahead.state.menuOpen).to.equal(false)
-        expect(typeahead.state.selected).to.equal(-1)
+        expect(typeahead.state.focused).to.equal(-1)
       })
 
       it('hides menu when query is empty', () => {
         typeahead.setState({ query: '' })
         typeahead.handleInputFocus()
         expect(typeahead.state.menuOpen).to.equal(false)
-        expect(typeahead.state.selected).to.equal(-1)
+        expect(typeahead.state.focused).to.equal(-1)
       })
     })
 
     describe('blurring input', () => {
       it('unfocuses component', () => {
-        typeahead.setState({ menuOpen: true, options: ['France'], selected: -1 })
+        typeahead.setState({ menuOpen: true, options: ['France'], focused: -1 })
         typeahead.handleInputBlur({ relatedTarget: null })
         expect(typeahead.state.menuOpen).to.equal(false)
-        expect(typeahead.state.selected).to.equal(null)
+        expect(typeahead.state.focused).to.equal(null)
       })
     })
 
@@ -132,51 +149,124 @@ describe('Typeahead', () => {
       it('sets the option as focused', () => {
         typeahead.setState({ options: ['France'] })
         typeahead.handleOptionFocus(0)
-        expect(typeahead.state.selected).to.equal(0)
+        expect(typeahead.state.focused).to.equal(0)
       })
     })
 
     describe('focusing out option', () => {
       it('unfocuses component', () => {
-        typeahead.setState({ menuOpen: true, options: ['France'], selected: 0 })
+        typeahead.setState({ menuOpen: true, options: ['France'], focused: 0 })
         typeahead.handleOptionFocusOut({ target: 'mock' }, 0)
         expect(typeahead.state.menuOpen).to.equal(false)
-        expect(typeahead.state.selected).to.equal(null)
+        expect(typeahead.state.focused).to.equal(null)
       })
     })
 
     describe('up key', () => {
       it('focuses previous element', () => {
-        typeahead.setState({ menuOpen: true, options: ['France'], selected: 0 })
+        typeahead.setState({ menuOpen: true, options: ['France'], focused: 0 })
         typeahead.handleKeyDown({ preventDefault: () => {}, keyCode: 38 })
-        expect(typeahead.state.selected).to.equal(-1)
+        expect(typeahead.state.focused).to.equal(-1)
       })
     })
 
     describe('down key', () => {
-      it('focuses next element', () => {
-        typeahead.setState({ menuOpen: true, options: ['France'], selected: -1 })
-        typeahead.handleKeyDown({ preventDefault: () => {}, keyCode: 40 })
-        expect(typeahead.state.selected).to.equal(0)
+      describe('0 options available', () => {
+        it('does nothing', () => {
+          typeahead.setState({ menuOpen: false, options: [], focused: -1 })
+          const stateBefore = typeahead.state
+          typeahead.handleKeyDown({ preventDefault: () => {}, keyCode: 40 })
+          expect(typeahead.state).to.equal(stateBefore)
+        })
+      })
+
+      describe('1 option available', () => {
+        it('focuses next element', () => {
+          typeahead.setState({ menuOpen: true, options: ['France'], focused: -1, selected: -1 })
+          typeahead.handleKeyDown({ preventDefault: () => {}, keyCode: 40 })
+          expect(typeahead.state.focused).to.equal(0)
+          expect(typeahead.state.selected).to.equal(0)
+        })
+      })
+
+      describe('2 or more option available', () => {
+        it('focuses next element', () => {
+          typeahead.setState({ menuOpen: true, options: ['France', 'Germany'], focused: 0, selected: 0 })
+          typeahead.handleKeyDown({ preventDefault: () => {}, keyCode: 40 })
+          expect(typeahead.state.focused).to.equal(1)
+          expect(typeahead.state.selected).to.equal(1)
+        })
+      })
+
+      describe('autoselect', () => {
+        describe('0 options available', () => {
+          it('does nothing', () => {
+            autoselectTypeahead.setState({ menuOpen: false, options: [], focused: -1, selected: -1 })
+            const stateBefore = autoselectTypeahead.state
+            autoselectTypeahead.handleKeyDown({ preventDefault: () => {}, keyCode: 40 })
+            expect(autoselectTypeahead.state).to.equal(stateBefore)
+          })
+        })
+
+        describe('1 option available', () => {
+          it('does nothing', () => {
+            autoselectTypeahead.setState({ menuOpen: true, options: ['France'], focused: -1, selected: 0 })
+            const stateBefore = autoselectTypeahead.state
+            autoselectTypeahead.handleKeyDown({ preventDefault: () => {}, keyCode: 40 })
+            expect(autoselectTypeahead.state).to.equal(stateBefore)
+          })
+        })
+
+        describe('2 or more option available', () => {
+          it('on input, focuses second element', () => {
+            autoselectTypeahead.setState({ menuOpen: true, options: ['France', 'Germany'], focused: -1, selected: 0 })
+            autoselectTypeahead.handleKeyDown({ preventDefault: () => {}, keyCode: 40 })
+            expect(autoselectTypeahead.state.focused).to.equal(1)
+            expect(autoselectTypeahead.state.selected).to.equal(1)
+          })
+        })
       })
     })
 
     describe('escape key', () => {
       it('unfocuses component', () => {
-        typeahead.setState({ menuOpen: true, options: ['France'], selected: -1 })
+        typeahead.setState({ menuOpen: true, options: ['France'], focused: -1 })
         typeahead.handleKeyDown({ preventDefault: () => {}, keyCode: 27 })
         expect(typeahead.state.menuOpen).to.equal(false)
-        expect(typeahead.state.selected).to.equal(null)
+        expect(typeahead.state.focused).to.equal(null)
       })
     })
 
     describe('enter key', () => {
-      it('closes the menu, sets the query, focuses the input', () => {
-        typeahead.setState({ menuOpen: true, options: ['France'], selected: 0 })
-        typeahead.handleKeyDown({ preventDefault: () => {}, keyCode: 13 })
-        expect(typeahead.state.menuOpen).to.equal(false)
-        expect(typeahead.state.query).to.equal('France')
-        expect(typeahead.state.selected).to.equal(-1)
+      describe('on an option', () => {
+        it('closes the menu, sets the query, focuses the input', () => {
+          typeahead.setState({ menuOpen: true, options: ['France'], focused: 0, selected: 0 })
+          typeahead.handleKeyDown({ preventDefault: () => {}, keyCode: 13 })
+          expect(typeahead.state.menuOpen).to.equal(false)
+          expect(typeahead.state.query).to.equal('France')
+          expect(typeahead.state.focused).to.equal(-1)
+          expect(typeahead.state.selected).to.equal(-1)
+        })
+      })
+
+      describe('on the input', () => {
+        it('does nothing', () => {
+          typeahead.setState({ menuOpen: true, options: ['France'], focused: -1, selected: -1 })
+          const stateBefore = typeahead.state
+          typeahead.handleKeyDown({ preventDefault: () => {}, keyCode: 13 })
+          expect(typeahead.state).to.equal(stateBefore)
+        })
+
+        describe('autoselect', () => {
+          it('closes the menu, selects the first option, keeps input focused', () => {
+            autoselectTypeahead.setState({ menuOpen: true, options: ['France'], focused: -1, selected: 0 })
+            autoselectTypeahead.handleKeyDown({ preventDefault: () => {}, keyCode: 13 })
+            expect(autoselectTypeahead.state.menuOpen).to.equal(false)
+            expect(autoselectTypeahead.state.query).to.equal('France')
+            expect(autoselectTypeahead.state.focused).to.equal(-1)
+            expect(autoselectTypeahead.state.selected).to.equal(-1)
+          })
+        })
       })
     })
   })